Create log.sh
Script file to create a job in interaction mode or submit a batch job on Triton. It creates an input folder, copies input files into input folder, reads the input files, then deletes the input folder when it is finished. It also prints out the progress of the MapReduce functions. The result is in a file called 'part-r-00000' located in the output directory.